pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eg4-monitor"
version = "0.1.0"
authors = [
  { name="John Hawkins", email="jthawkins2001@gmail.com" },
]
description = "A Python library to monitor EG4 Solar Inverters via the EG4 Monitor website."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Home Automation",
]
dependencies = [
    "requests>=2.25.0",
    "pydantic>=2.0.0",
    "httpx>=0.20.0",
    "python-dotenv>=0.19.0",
]

[project.scripts]
eg4-monitor = "eg4_monitor.cli:main"

[project.urls]
"Homepage" = "https://github.com/beardface/eg4-ctrl"
"Bug Tracker" = "https://github.com/beardface/eg4-ctrl/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["eg4_monitor*"]
